Add files via upload
First commit, was working until I added the dao and service classes. Now it says that I am missing an AutoWired annotation in the dao on the SessionFactory, but it is there. I need to fix that and then I just need to add methods to the controller for the crud functionality.